requirements.txt
fake-useragent~=1.4.0
SQLAlchemy~=2.0.23
requests~=2.31.0
beautifulsoup4~=4.12.2
psycopg2~=2.9.9
python-dotenv~=1.0.0